requirements.txt
boto3==1.23.1
python-dotenv==0.21.0
requests==2.28.1
web3==5.31.3